Fix Manager errors

+++ gestor-gastos-web/src/app/access.guard.ts
@@ -1,9 +1,7 @@
 import { Injectable } from '@angular/core';
 import {
   ActivatedRouteSnapshot,
-  CanActivate,
-  RouterStateSnapshot,
-  UrlTree,
+  CanActivate
 } from '@angular/router';
 import { Observable } from 'rxjs';
 import { ApiConnectionService } from './api-connection.service';
@@ -18,24 +16,15 @@ export class AccessGuard implements CanActivate {
     route: ActivatedRouteSnapshot
   ): Observable<boolean> | Promise<boolean> | boolean {
     const requiresLogin = route.data['requiresLogin'] || false;
-    if (requiresLogin) {
-      return ApiConnectionService.isApiAlive().then((response) => {
-        if (response != undefined) {
-          this.localStorageService.set('apiAlive', true);
-          return ApiConnectionService.isTokenAvailable().then((response) => {
-            if ('token_info' in response) return true;
-            else {
-              this.localStorageService.remove('token');
-              this.localStorageService.remove('username');
-              return false;
-            }
-          });
-        }
-        else {
-          this.localStorageService.set('apiAlive', false);
-          return false;
-        }
-      });
-    } else return true;
+    return ApiConnectionService.isTokenAvailable().then((response) => {
+      if (response != undefined && 'token_info' in response) return true;
+      else {
+        this.localStorageService.remove('token');
+        this.localStorageService.remove('username');
+        if (requiresLogin) return false;
+        else return true;
+      }
+    });
+    
   }
 }
